Jiminy 1.4 (#207)
* [Gym] Generic locomotion envs - PART I (#194)

* [Viewer] Replace multi args "," delimiter by "|" to allow comma in path. Raise exception is
* [Gym] Major rework to add base locomotion learning environments.
* [Core] Implement methods to add and remove frames from the robot model.
* [Core] Add 'mechanicalReduction' option to the motors.
* [Core] Add support of mesh collisions using HPP-FCL. Migration to pinocchio/eigenpy v2.5.0.
* [misc] Cleanup cmake config file to implement actual legacy mode for Ubuntu18. Update Ubuntu 18 easy install deps.
* [misc] Add official support of Ubuntu 20.
* [Core] Add support of body collision with the ground, but do NOT actually compute the forces.


* [Gym] Generic locomotion envs - PART II (#196)

* [core] Expose collision model/data in Python. Fix mesh path at robot level instead of viewer.
* [core] Make clear the difference between body collisions and contact frames. Implement actual force sensor.
* [core] Fix collision computation. Add option for max collision contact points.
* [core] Various minor improvements. Enable FCL convex hull computation.
* [core] Implement collision body contact force. Use convex hull instead of the actual geometry for evaluating the collisions.
* [misc] Fix compilation flags on windows. Update Pinocchio to the latest available release on Ubuntu 18 (same as 20).
* [Viewer] Remove deprecation warning of Pinocchio. Do not try to load plane primitive associated with the ground.
* [Viewer] Minor improvement of error handling.


* [Gym] Generic locomotion envs - PART III (#197)

* [core] Add new 'successiveIterFailedMax' stepper option.
* [core] Fix collision body contact force computation and primitive geometry support. Fix add frame. Fix generic robot envs.
* [core] Improve mesh package support. Add meshPackageDirs optional argument to model.initialize method. 
* [core] Expose sensor data individually. Add sensor __repr__. Expose EngineMultiRobot systems vector in Python. Fix systemData returning unusable robot and controller.
* [engine_async] Improve viewer backend management.
* [unit_py] Rework the unit test in Python to reduce code duplication. Add dedicated box and sphere testing urdf with collision and visual.
* [misc] Reduce static library compile size by removing debugging symbols.

* [python/robot] Set controller and sensors update periods.


* [Gym] Generic locomotion envs - PART IV (#198)

* [core] Add ControllerFunctor default value 'None' in Python. Take advantage of it for EngineAsync. 
* [python] Improve PEP8 compliance and typing.


* [Gym] Generic locomotion envs - PART V (#199)

* [python] Remove Python2 support for Python bindings.
* [core] Fix exception catching addCouplingForce.
* [gym] Use typing.
* [gym] Add generic locomotion env.


* [Gym] Generic locomotion envs - PART VI (#200)

* [core] More specific Utilities error messages.
* [core] Add support of collision with bodies associated with multiple geometries.
* [misc] Multi-level data folder architecture.
* [misc] Rename gym env_base.py in env_bases.py.
* [python] Enable to specify collision bodies, contact points, and ground model in TOML config file.
* [python] Add IMU sensor to root body by default systematically if no Gazebo IMU has been found.
* [python] BaseJiminyEngine now creates a temporary toml file if not available.
* [gym] Enable to refresh observation and action spaces separately. Enable to use custom PID gains.
* [gym] Remove 'TorqueControl' from walker name since it is the default.


* [Gym] Generic locomotion envs - PART VII (#201)

* [misc] Remove Python 2 compatibility code.
* [core] Fix Model 'removeCollisionBodies' method.
* [core] Fix support of multiple geometry by body for collision computation.
* [core] Compute the centers of mass systematically.
* [core] Check that the initial state is not out of bounds.
* [core|gym] Properly handle step integration failure.
* [gym] Fix 'done' flag handling in 'step' method. Remove 'engine' property to avoid redundancy.
* [gym] Use sample_state to get initial height.
* [gym] Fix PD Control env motor-to-encoder mapping.


* [Gym] Generic locomotion envs - PART VIII (#202)

* [misc] Move meshes in subdirectories.
* [core] Add support of 'continuous' joint type.
* [core] Check that at least one geometry is associated with the bodies when adding collisions.
* [core] Fix AddFrame to check not already exists, and update pinocchio data.
* [core] Add optional parameter to ignore meshes when adding collision body.
* [core] Fix support of mechanical reduction.
* [python] Add support of encoders and motors of any 1-dof type. Better handling of default sensor and controller update period.
* [python] Fix force sensor and collision guess if no gazebo plugin available.
* [python] Improve handling of collision bodies by adding several fallback mechanisms.
* [python] Make sure that up to 3 points are in contact, instead of being aligned with the closest contact point frame.
* [python] Fix motor inertia and mechanical reduction not properly extracted from the URDF.
* [gym] Apply ground offset at 'reset' instead of '_sample_state'.


* [Gym] Generic locomotion envs - PART IX (#203)

* [misc] Update dependency versions.
* [core] Minor rework for engine forward dynamics computation.
* [python/dynamics] Fix dynamics quantities not properly initialized at engine reset.
* [python/dynamics] Take into account collision body orientation for box primitive when computing freeflyer.
* [python/robot] Automatically convert box primitive body collision in contact points for numerical robustness.
* [python/robot] Improve default toml generation.
* [python/Viewer] Fix play_trajectory 'wait_for_client' option not used.
* [python/Viewer] Refactor Viewer to make it simpler and clearer.
* [python/Viewer] Use typing everywhere in Viewer.
* [gym] Enable to define custom neutral configuration.


* [Gym] Generic locomotion envs - PART X (#204)

* [core] Fix various segfault in case of exceptions. More verbose error message when failing to load URDF model.
* [core] Enable to postpone Controller def after engine init, and to change it between simu.
* [python] Remove meshes and use primitives for toys models.
* [python] Refactor of engine/simulator classes to be more versatile yet simpler.
* [python/Viewer] More robust handling of meshcat comms in Jupyter.


* Generic locomotion envs (#206)

* [python/simulator] Add 'plot' method to simulator to display simulation data up to current time.
* [python/examples] Remove messy examples and fix the others.
* [gym] Minor refactoring to enable getting normal env from goal env without being hacky nor code duplication.
* [gym] Enable to set camera pose in env 'render' method. Remove 'is_success' info in base env.
* [gym] Fix terminal reward computation.


* [core] Implement Lie-group based steppers. (#205)

* [core] Implement custom steppers working in Lie group directly, for improved accuracy and performance.
* [core] Add constraint for wheel rolling on flat ground.


* Bug fixes and stepper performance improvement. (#208)

* [msic] Various bug fixes.
* [misc] Enable Eigen 3.3 for any plateform since it is no longer conflicting with Boost.
* [core] Use enable_if_t from STL C++14.
* [core] Rework Lie Algebra implementation for efficiency.
* [core] Stepper performance improvement by using buffers.
* [misc] Fix 'set-env' github CI deprecations.

+++ python/jiminy_py/src/jiminy_py/meshcat/recorder.py
@@ -10,8 +10,9 @@ import subprocess
 import multiprocessing
 import multiprocessing.managers
 from pathlib import Path
-from ctypes import c_char_p, c_bool, c_int
+from ctypes import c_char_p
 from contextlib import redirect_stderr
+from typing import Awaitable, Any, Optional
 
 from IPython import get_ipython
 shell = get_ipython().__class__.__module__
@@ -23,7 +24,7 @@ if shell.startswith('google.colab.'):
     pyppeteer.chromium_downloader.chromium_executable = \
         lambda : Path("/usr/lib/chromium-browser/chromium-browser")
     if not pyppeteer.chromium_downloader.check_chromium():
-        logging.warning("Chrome must be installed manually on Google Colab. "\
+        logging.warning("Chrome must be installed manually on Google Colab. "
             "It must be done using '!apt install chromium-chromedriver'.")
 else:
     # Must use a recent release that supports webgl rendering with hardware
@@ -34,7 +35,7 @@ else:
 from pyppeteer.connection import Connection
 from pyppeteer.browser import Browser
 from pyppeteer.launcher import Launcher, get_ws_endpoint
-from requests_html import HTMLSession
+from requests_html import HTMLSession, HTMLResponse
 
 # ================ Monkey-patch =======================
 
@@ -88,7 +89,9 @@ Launcher.launch = launch
 
 # ======================================================
 
-async def capture_frame_async(client, width, height):
+async def capture_frame_async(client: HTMLResponse,
+                              width: int,
+                              height: int) -> Awaitable[Any]:
     """
     @brief    Send a javascript command to the hidden browser to
               capture frame, then wait for it (since it is async).
@@ -108,7 +111,10 @@ async def capture_frame_async(client, width, height):
         }
     """)
 
-async def start_video_recording_async(client, fps, width, height):
+async def start_video_recording_async(client: HTMLResponse,
+                                      fps: int,
+                                      width: int,
+                                      height: int) -> Awaitable[None]:
     await client.html.page.setViewport(
             {'width': width, 'height': height})
     await client.html.page.evaluate(f"""
@@ -121,7 +127,7 @@ async def start_video_recording_async(client, fps, width, height):
         }}
     """)
 
-async def add_video_frame_async(client):
+async def add_video_frame_async(client: HTMLResponse) -> Awaitable[None]:
     await client.html.page.evaluate("""
         () => {
             viewer.renderer.render(viewer.scene, viewer.camera);
@@ -129,7 +135,8 @@ async def add_video_frame_async(client):
         }
     """)
 
-async def stop_and_save_video_async(client, path):
+async def stop_and_save_video_async(client: HTMLResponse,
+                                    path: str) -> Awaitable[None]:
     directory = os.path.dirname(path)
     filename = os.path.splitext(os.path.basename(path))[0]
     await client.html.page._client.send('Page.setDownloadBehavior',
@@ -146,7 +153,9 @@ async def stop_and_save_video_async(client, path):
         }}
     """)
 
-def meshcat_recorder(meshcat_url, request_shm, message_shm):
+def meshcat_recorder(meshcat_url: str,
+                     request_shm: multiprocessing.Value,
+                     message_shm: multiprocessing.Value) -> None:
     # Do not catch signal interrupt automatically, to avoid killing meshcat
     # server and stopping Jupyter notebook cell.
     signal.signal(signal.SIGINT, signal.SIG_IGN)
@@ -161,7 +170,8 @@ def meshcat_recorder(meshcat_url, request_shm, message_shm):
     client.html.render(keep_page=True)
     message_shm.value = f"{session._browser.process.pid}"
 
-    # Stop the animation loop by default, since it is not relevant for recording only
+    # Stop the animation loop by default, since it is not relevant for
+    # recording only
     async def stop_animate_async(client):
         await client.html.page.evaluate("""
             () => {
@@ -177,7 +187,7 @@ def meshcat_recorder(meshcat_url, request_shm, message_shm):
                 while request_shm.value != "quit":  # [>Python3.8] while (request := request_shm.value) != "quit":
                     request = request_shm.value
                     if request != "":
-                        args = map(str.strip, message_shm.value.split(","))
+                        args = map(str.strip, message_shm.value.split("|"))
                         if request == "take_snapshot":
                             width, height = map(int, args)
                             coro = capture_frame_async(client, width, height)
@@ -219,7 +229,7 @@ class MeshcatRecorder:
               Process to enable parallel asyncio loop execution, which
               is necessary to support recording in Jupyter notebook.
     """
-    def __init__(self, url):
+    def __init__(self, url: str):
         self.is_open = False
         self.is_recording = False
         self.url = url
@@ -228,7 +238,7 @@ class MeshcatRecorder:
         self.proc = None
         self.__browser_pid = None
 
-    def open(self):
+    def open(self) -> None:
         self.__manager = multiprocessing.managers.SyncManager()
         self.__manager.start(
             lambda : signal.signal(signal.SIGINT, signal.SIG_IGN))
@@ -254,7 +264,7 @@ class MeshcatRecorder:
     def __del__(self):
         self.release()
 
-    def release(self):
+    def release(self) -> None:
         if self.__shm is not None:
             if self.proc.is_alive():
                 try:
@@ -278,7 +288,10 @@ class MeshcatRecorder:
             self.__manager = None
         self.is_open = False
 
-    def _send_request(self, request, message=None, timeout=2.0):
+    def _send_request(self,
+                      request: str,
+                      message: Optional[str] = None,
+                      timeout: float = 2.0) -> None:
         if not self.is_open:
             raise RuntimeError(
                 "Meshcat recorder is not open. Impossible to send requests.")
@@ -295,28 +308,32 @@ class MeshcatRecorder:
             elif not self.proc.is_alive():
                 self.release()
                 raise RuntimeError(
-                    "Backend browser has encountered an unrecoverable "\
+                    "Backend browser has encountered an unrecoverable "
                     "error: ", self.__shm['message'].value)
 
-    def capture_frame(self, width=None, height=None):
+    def capture_frame(self,
+                      width: Optional[int] = None,
+                      height: Optional[int] = None) -> str:
         self._send_request("take_snapshot",
-            message=f"{width if width is not None else -1},"\
-                    f"{height if height is not None else -1}")
+            message=(f"{width if width is not None else -1}|"
+                     f"{height if height is not None else -1}"))
         return self.__shm['message'].value
 
-    def start_video_recording(self, fps, width, height):
+    def start_video_recording(self,
+                              fps: float,
+                              width: int,
+                              height: int) -> None:
         self._send_request("start_record",
-            message=f"{fps},{width},{height}", timeout=10.0)
+            message=f"{fps}|{width}|{height}", timeout=10.0)
         self.is_recording = True
 
-    def add_video_frame(self):
+    def add_video_frame(self) -> None:
         if not self.is_recording:
-            raise RuntimeError(
-                "No video being recorded at the moment. "\
+            raise RuntimeError("No video being recorded at the moment. "
                 "Please start recording before adding frames.")
         self._send_request("add_frame")
 
-    def stop_and_save_video(self, path):
+    def stop_and_save_video(self, path: str) -> bool:
         def file_available(path):
             if not os.path.exists(path):
                 return False
@@ -332,9 +349,11 @@ class MeshcatRecorder:
             return True
 
         if not self.is_recording:
-            raise RuntimeError(
-                "No video being recorded at the moment. "\
+            raise RuntimeError("No video being recorded at the moment. "
                 "Please start recording and add frames before saving.")
+        if "|" in path:
+            raise ValueError(
+                "'|' character is not supported in video export path.")
         path = os.path.abspath(pathlib.Path(path).with_suffix('.webm'))
         if os.path.exists(path):
             os.remove(path)
